+++ R/filter_by_key.R
@@ -40,41 +40,35 @@ filter_by_key<- function(dframe,
                          ...,
                          select = NULL) {
 
+  # force arguments after dots to be used by name
   wrapr::stop_if_dot_args(
     substitute(list(...)), "Arguments after ... must be named,"
   )
 
-  # default optional arguments
+  # explicit arguments
+  assert_explicit(dframe)
+  assert_explicit(match_to)
+  assert_explicit(key_col)
+
+  # check argument class
+  # set NULL default once dframe class checked
   assert_class(dframe, "data.frame")
   assert_class(match_to, "data.frame")
-  select <- select %||% names(dframe)
-
-  # check arguments
   assert_class(key_col, "character")
+  select <- select %||% names(dframe)
   assert_class(select, "character")
 
   # existence of required columns
-  assert_required_column(dframe, key_col)
-  assert_required_column(match_to, key_col)
+  assert_common_column(dframe, key_col)
+  assert_common_column(match_to, key_col)
 
-  # to use DT syntax in next assertions
-  setDT(dframe)
+  # dframe is not modified by reference
+  dframe <- copy(as.data.table(dframe))
   setDT(match_to)
-  dframe <- copy(dframe)
-
-  # class of required columns
-  # assert_class(dframe[get(key_col)], "character")
-  # assert_class(match_to[get(key_col)], "character")
-
-  # bind names due to nonstandard evaluation notes in R CMD check
-  # var <- NULL
-
 
   # key must be in both data frames
   stopifnot("`key_col` must be a column in `dframe`" = key_col %in% names(dframe))
   stopifnot("`key_col` must be a column in `match_to`" = key_col %in% names(match_to))
-
-  if (is.null(select)) {select <- names(dframe)}
 
   # keep one column
   match_to <- match_to[, key_col, with = FALSE]